lang: TempoClock: Ensure that thread seconds are always increasing
In TempoClock, this scheduled item's beats must be >= the last one.
And this item's seconds are always >= previous.

In LinkClock, when the tempo changes, beats(n) >= beats(n-1)
but seconds(n) may sometimes be < seconds(n-1).

We detect this case and substitute an alternate scheduling seconds,
derived from previous seconds + (deltaBeats * beatDur).